pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyastro"
version = "0.0.0"
description = "Простой проект астрологических вычислений с Swiss Ephemeris"
authors = [
    { name = "Nikolai Pakulin" }
]
requires-python = ">=3.12"

dependencies = [
    "pyswisseph>=2.10",
    "pyyaml"
]

[project.scripts]
pyastro = "pyastro.main:main"
pyastro-parade = "pyastro.parade:main"
pyastro-validate = "pyastro.validate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pylint."MESSAGES CONTROL"]
disable = "line-too-long,no-else-return,import-outside-toplevel,superfluous-parens"